Update basic info, layout, and functionality of homepage buttons
