+++ app/index.tsx
@@ -28,29 +28,6 @@ export default function Index() {
         <DynamicEventSection></DynamicEventSection>
         <ContactsIcons></ContactsIcons> 
         <Resources></Resources>
-        <DynamicButton buttons={[
-            {
-                type: 2, // internal - forms/calendar
-                shape: 0, // full
-                buttonConfig: {
-                    text: "Calendar",
-                    icon: "calendar",
-                    internalLink: `../(tabs)/home/calendar`,
-                    backgroundColor: "rgba(255,255,255,1)"
-                 }
-            },
-            {
-                type: 2, // internal - forms/calendar
-                shape: 0, // full
-                buttonConfig: {
-                    text: "Forms",
-                    icon: "edit",
-                    internalLink: `../(tabs)/home/forms`,
-                    backgroundColor: "rgba(255,255,255,1)"
-                 }
-            }
-          ]
-        }/>
       </View>
     </ScrollView>
   );
