+++ app/lyrics.tsx
@@ -1,0 +1,264 @@
+import { ScrollView, Text, View, StyleSheet } from 'react-native';
+
+const LyricsPage = () => {
+  return (
+    <ScrollView contentContainerStyle={styles.container}>
+      <Text style={styles.heading}>Worship Lyrics</Text>
+
+      <Text style={styles.song_title}>Only wanna sing</Text>
+
+      <View style={styles.lyricsBlock}>
+        <Text style={styles.lyrics}>
+          This is no performance{'\n'}
+          Lord, I pray it's worship{'\n'}
+          Empty words I can't afford{'\n'}
+          I'm not chasing feelings{'\n'}
+          That's not why I'm singing{'\n'}
+          You're the reason for my song{'\n\n'}
+          
+          I only wanna sing{'\n'}
+          If I sing with everything{'\n'}
+          If I sing for You my King{'\n'}
+          I can't imagine why{'\n'}
+          I would do this all for hype{'\n'}
+          'Cause it's all to lift You high{'\n\n'}
+
+          You don't want perfection{'\n'}
+          Just my soul's attention{'\n'}
+          All I have is what I'll give{'\n\n'}
+
+          More than a song that lasts a moment{'\n'}
+          I'll live a life full of honest worship{'\n'}
+          If I'm here to sing then I'll sing with purpose{'\n'}
+          All the praise, Lord, You deserve it, now{'\n\n'}
+
+          I only wanna sing{'\n'}
+          If I sing with everything{'\n'}
+          If I sing for You my King{'\n'}
+          I can't imagine why{'\n'}
+          I would do this all for hype{'\n'}
+          'Cause it's all to lift You high{'\n\n'}
+
+          I only wanna sing{'\n'}
+          If I sing with everything{'\n'}
+          If I sing for You my King{'\n'}
+        </Text>
+
+        <Text style={styles.song_title}>Такий Великий Бог</Text>
+
+        <Text style={styles.lyrics}>
+          Нехай поклоняться Тобі, всі народи, Бог мій,{'\n'}
+          Бо Ти Один достойний слави, Ти Один Святий{'\n'}
+          Нехай співають всі Тобі Слава , слава , слава ,{'\n'}
+          Слава , слава , слава Тобі{'\n\n'}
+
+          Нехай молитви всі Тобі, з подякою складають{'\n'}
+          Бо Ти Один достойний слави, Добрий тільки Ти{'\n'}
+          Нехай співають всі Тобі Слава , слава, слава ,{'\n'}
+          Слава , слава, слава Тобі{'\n\n'}
+
+          Такий Великий  Бог, а так близько{'\n'}
+          Ти так близько , в серці моїм{'\n'}
+          Хоч небо над землею, так високо{'\n'}
+          А ти близько, в серці моїм{'\n\n'}
+
+          Нехай хвала із наших уст, буде Тобі приємна{'\n'}
+          Господь Один достойний слави, Ти Один Святий{'\n'}
+          Нехай у єдності серця співають Слава, слава,{'\n'}
+          Слава , слава, слава Тобі{'\n\n'}
+
+          Нехай у радості земля Спаситель наш чекає,{'\n'}
+          Господь Один достойний слави, Добрий тільки Ти{'\n'}
+          Нехай співають всі Тобі Слава, слава, слава,{'\n'}
+          Слава, слава, слава Тобі{'\n\n'}
+
+          Такий Великий  Бог, а так близько{'\n'}
+          Ти так близько , в серці моїм{'\n'}
+          Хоч небо над землею, так високо{'\n'}
+          А ти близько, в серці моїм{'\n\n'}
+
+          Такий Великий  Бог, а так близько{'\n'}
+          Ти так близько , в серці моїм{'\n'}
+          Хоч небо над землею, так високо{'\n'}
+          А ти близько, в серці моїм{'\n'}
+        </Text>
+
+        <Text style={styles.song_title}>Goodness of God</Text>
+
+        <Text style={styles.lyrics}>
+          I love You, Lord{'\n'}
+          For Your mercy never fails me{'\n'}
+          All my days, I've been held in Your hands{'\n'}
+          From the moment that I wake up{'\n'}
+          Until I lay my head{'\n'}
+          Oh, I will sing of the goodness of God{'\n\n'}
+
+          'Cause all my life You have been faithful{'\n'}
+          And all my life You have been so, so good{'\n'}
+          With every breath that I am able{'\n'}
+          Oh, I will sing of the goodness of God{'\n\n'}
+
+          I love Your voice{'\n'}
+          You have led me through the fire{'\n'}
+          In darkest night You are close like no other{'\n'}
+          I've known You as a Father{'\n'}
+          I've known You as a Friend{'\n'}
+          And I have lived in the goodness of God, yeah{'\n\n'}
+
+          'Cause all my life You have been faithful, oh yes You have{'\n'}
+          And all my life You have been so, so good{'\n'}
+          With every breath that I am able{'\n'}
+          Oh, I will sing of the goodness of God{'\n\n'}
+
+          Your goodness is running after, it's running after me{'\n'}
+          Your goodness is running after, it's running after me{'\n'}
+          With my life laid down, I surrendered now{'\n'}
+          I give You everything, oh Lord{'\n'}
+          Your goodness is running after, it's running after me Your goodness is running after, it's{'\n'}
+          running after me{'\n\n'}
+
+          Your goodness is running after, it's running after me{'\n'}
+          With my life laid down, I surrendered now{'\n'}
+          I give You everything, oh Lord{'\n'}
+          Your goodness is running after, it's running after me{'\n\n'}
+
+          And all my life You have been faithful{'\n'}
+          And all my life You have been so, so good{'\n'}
+          With every breath that I am able{'\n'}
+          Oh, I'm gonna sing of the goodness of God{'\n'}
+        </Text>
+
+        <Text style={styles.song_title}>Ісус, Ти переміг</Text>
+
+        <Text style={styles.lyrics}>
+          Невидиму зброю нам дав{'\n'}
+          Щоб ворог схилився і впав{'\n'}
+          Коли славлю Тебе, Бог{'\n'}
+          Коли є навколо біда{'\n'}
+          Прийде перемога Твоя{'\n'}
+          Коли славлю Тебе, Бог{'\n\n'}
+
+          Ісус, Ти переміг{'\n'}
+          Весь світ у Твоїх ніг{'\n'}
+          Ти правиш навіки{'\n'}
+          І не встоїть ніхто{'\n'}
+          Перед Тобою, Бог{'\n'}
+          Твоє Царство навіки{'\n\n'}
+
+          Ти наміри пекла розбив{'\n'}
+          Ніщо не заглушить хвали{'\n'}
+          Коли славлю Тебе, Бог{'\n'}
+          Зруйнуєш Ти стіни гріха{'\n'}
+          І прийде свобода Твоя{'\n'}
+          Коли славлю Тебе, Бог{'\n\n'}
+
+          Ісус, Ти переміг{'\n'}
+          Весь світ у Твоїх ніг{'\n'}
+          Ти правиш навіки{'\n'}
+          І не встоїть ніхто{'\n'}
+          Перед Тобою, Бог{'\n'}
+          Твоє Царство навіки{'\n\n'}
+
+          Звучить Твій голос неначе грім{'\n'}
+          О, Ти сильний Бог, Ти сильний Бог{'\n'}
+          І як на Небі лунає спів{'\n'}
+          О, Ти сильний Бог, Ти сильний Бог{'\n\n'}
+
+          Звучить Твій голос неначе грім{'\n'}
+          О, Ти сильний Бог, Ти сильний Бог{'\n'}
+          І як на Небі лунає спів{'\n'}
+          О, Ти сильний Бог, Ти сильний Бог{'\n\n'}
+
+          Ісус, Ти переміг{'\n'}
+          Весь світ у Твоїх ніг{'\n'}
+          Ти правиш навіки{'\n'}
+          І не встоїть ніхто{'\n'}
+          Перед Тобою, Бог{'\n'}
+          Твоє Царство навіки{'\n'}
+        </Text>
+
+        <Text style={styles.song_title}>Все що я прагну</Text>
+
+        <Text style={styles.lyrics}>
+          Ти життя відав на хресті за мій гріх{'\n'}
+          На вівтар Свою кров пролив і спас всіх{'\n'}
+          Ім'я моє записав в небесах{'\n'}
+          Ти для мене все,{'\n'}
+          Ти для мене все,{'\n\n'}
+
+          Все що я прагну{'\n'}
+          Поруч бути і Тебе знати{'\n'}
+          Моє серце назавжди Ти змінив{'\n'}
+          І душа моя співає Тобі{'\n\n'}
+
+          Все що я прагну{'\n'}
+          Поруч бути і Тебе знати{'\n'}
+          Моє серце назавжди Ти змінив{'\n'}
+          І душа моя співає Тобі{'\n\n'}
+
+          Ти життя відав на хресті за мій гріх{'\n'}
+          На вівтар Свою кров пролив і спас всіх{'\n'}
+          Ім'я моє записав в небесах{'\n'}
+          Ти для мене все,{'\n'}
+          Ти для мене все,{'\n\n'}
+
+          Все що я прагну{'\n'}
+          Поруч бути і Тебе знати{'\n'}
+          Моє серце назавжди Ти змінив{'\n'}
+          І душа моя співає Тобі{'\n\n'}
+
+          Все що я прагну{'\n'}
+          Поруч бути і Тебе знати{'\n'}
+          Моє серце назавжди Ти змінив{'\n'}
+          І душа моя співає Тобі{'\n\n'}
+
+          Я в присутності Твоїй руки підіймаю{'\n'}
+          Ти в серці моїм{'\n'}
+          Від душі Тебе Господь я Благословляю{'\n'}
+          У всьому є Ти{'\n'}
+          Я в присутності Твоїй руки підіймаю{'\n'}
+          Ти в серці моїм{'\n'}
+          Від душі Тебе Господь я Благословляю{'\n'}
+          У всьому є Ти{'\n\n'}
+
+          Все що я прагну{'\n'}
+          Поруч бути і Тебе знати{'\n'}
+          Моє серце назавжди Ти змінив{'\n'}
+          І душа моя співає Тобі{'\n\n'}
+
+          Все що я прагну{'\n'}
+          Поруч бути і Тебе знати{'\n'}
+          Моє серце назавжди Ти змінив{'\n'}
+          І душа моя співає Тобі{'\n\n'}
+        </Text>
+      </View>
+    </ScrollView>
+  );
+};
+
+const styles = StyleSheet.create({
+  container: {
+    padding: 20,
+    backgroundColor: '#fff',
+  },
+  heading: {
+    fontSize: 20,
+    fontWeight: 'bold',
+    marginBottom: 20,
+    textAlign: 'center',
+  },
+  lyricsBlock: {
+    marginBottom: 10,
+  },
+  lyrics: {
+    fontSize: 16,
+    lineHeight: 24,
+  },
+  song_title: {
+    fontSize: 20,
+    fontWeight: 'bold',
+    marginBottom: 20,
+  },
+});
+
+export default LyricsPage;
