+++ app/solMission.tsx
@@ -1,0 +1,10 @@
+import { Text } from "react-native";
+
+
+const SOLMission = () => {
+    return (
+        <Text>SOL Mission Page</Text>
+    );
+};
+
+export default SOLMission;
