+++ app/solRu.tsx
@@ -1,0 +1,10 @@
+import { Text } from "react-native";
+
+
+const SOLru = () => {
+    return (
+        <Text>SOL Ru Page</Text>
+    );
+};
+
+export default SOLru;
